Remove ImageInternal.swift
